changed Int with Double

+++ data/src/main/java/dev/vincenzocostagliola/data/domain/CoinData.kt
@@ -100,7 +100,7 @@ data class CoinData(
         )
 
         data class MarketCap(
-            val eur: Long
+            val eur: Double
         )
     }
 }
